feat(config): allow loading config from external config file (#77)
* feat(config): allow loading config from external config file

* feat(config): adds additional package for config loading

* fix(config): removes modulename config

* chore: remove deprecated lines in husky scripts

* fix(config): use contentful config key for fakes & typings

+++ packages/contentful-ssg/src/cli.ts
@@ -37,11 +37,15 @@ const parseFetchArgs = (cmd: {
   verbose: boolean;
   ignoreErrors: boolean;
   query: string;
-}): Partial<Config> => ({
+  cwd?: string;
+  config?: string;
+}): Partial<Config> & { cwd?: string; configFile?: string } => ({
   preview: cmd.preview,
   verbose: cmd.verbose,
   ignoreErrors: cmd.ignoreErrors,
   query: parseQuery(cmd.query),
+  cwd: cmd?.cwd || process.cwd(),
+  configFile: cmd?.config,
 });
 
 type CommandError = Error & {
@@ -71,21 +75,25 @@ program
   .command('init')
   .description('Initialize contentful-ssg')
   .option('--typescript', 'Initialize typescript config')
+  .option(
+    '--config <configFile>',
+    'Use this configuration, overriding other config options if present',
+  )
+  .option('--cwd <directory>', 'Working directory. Defaults to process.cwd()')
   .action(
-    actionRunner(async (cmd) => {
+    actionRunner(async (cmd: { typescript?: boolean; cwd?: string; config?: string }) => {
       const useTypescript = Boolean(cmd?.typescript ?? false);
-      const config = await getConfig();
-      const verified = await askAll(config);
+      const cwd = cmd?.cwd ?? process.cwd();
+      const configFile = cmd?.config;
+      const config = await getConfig({ cwd, configFile });
 
       const environmentConfig = getEnvironmentConfig();
 
-      const filePath = path.join(
-        process.cwd(),
-        `contentful-ssg.config.${useTypescript ? 'ts' : 'js'}`,
-      );
+      const filePath =
+        configFile || path.join(cwd, `contentful-ssg.config.${useTypescript ? 'ts' : 'js'}`);
       const prettierOptions = await prettier.resolveConfig(filePath);
-      if (verified.directory?.startsWith('/')) {
-        verified.directory = path.relative(process.cwd(), verified.directory);
+      if (config.directory?.startsWith('/')) {
+        config.directory = path.relative(cwd, config.directory);
       }
 
       const environmentKeys: Array<keyof ContentfulConfig> = Object.keys(
@@ -96,29 +104,29 @@ program
       if (environmentConfig && existsSync('.env')) {
         const envSource = await readFile('.env', 'utf8');
         const nextEnvSource = envSource
-          .replace(/(CONTENTFUL_SPACE_ID\s*=\s*['"]?)[^'"]*(['"]?)/, `$1${verified.spaceId}$2`)
+          .replace(/(CONTENTFUL_SPACE_ID\s*=\s*['"]?)[^'"]*(['"]?)/, `$1${config.spaceId}$2`)
           .replace(
             /(CONTENTFUL_ENVIRONMENT_ID\s*=\s*['"]?)[^'"]*(['"]?)/,
-            `$1${verified.environmentId}$2`,
+            `$1${config.environmentId}$2`,
           )
           .replace(
             /(CONTENTFUL_MANAGEMENT_TOKEN\s*=\s*['"]?)[^'"]*(['"]?)/,
-            `$1${verified.managementToken}$2`,
+            `$1${config.managementToken}$2`,
           )
           .replace(
             /(CONTENTFUL_PREVIEW_TOKEN\s*=\s*['"]?)[^'"]*(['"]?)/,
-            `$1${verified.previewAccessToken}$2`,
+            `$1${config.previewAccessToken}$2`,
           )
           .replace(
             /(CONTENTFUL_DELIVERY_TOKEN\s*=\s*['"]?)[^'"]*(['"]?)/,
-            `$1${verified.accessToken}$2`,
+            `$1${config.accessToken}$2`,
           );
 
         await outputFile('.env', nextEnvSource);
       }
 
       const cleanedConfig = omitKeys(
-        verified,
+        config,
         'preview',
         'verbose',
         'rootDir',
@@ -129,7 +137,7 @@ program
       );
 
       let content = '';
-      if (useTypescript) {
+      if (useTypescript || filePath.endsWith('.ts')) {
         content = await prettier.format(
           `import {Config} from '@jungvonmatt/contentful-ssg';
         export default <Config>${JSON.stringify(cleanedConfig)}`,
@@ -138,27 +146,30 @@ program
             ...prettierOptions,
           },
         );
-      } else {
+      } else if (filePath.endsWith('.js')) {
         content = await prettier.format(`module.exports = ${JSON.stringify(cleanedConfig)}`, {
           parser: 'babel',
           ...prettierOptions,
         });
       }
 
-      let writeFile = true;
-      if (existsSync(filePath)) {
-        writeFile = await confirm(
-          `Config file already exists. Overwrite?\n\n${chalk.reset(content)}`,
-        );
-      } else {
-        writeFile = await confirm(`Please verify your settings:\n\n${chalk.reset(content)}`, true);
-      }
+      if (filePath.endsWith('.js') || filePath.endsWith('.ts')) {
+        let writeFile = true;
+        if (existsSync(filePath)) {
+          writeFile = await confirm(
+            `Config file already exists. Overwrite?\n\n${chalk.reset(content)}`,
+          );
+        } else {
+          writeFile = await confirm(
+            `Please verify your settings:\n\n${chalk.reset(content)}`,
+            true,
+          );
+        }
 
-      if (writeFile) {
-        await outputFile(filePath, content);
-        console.log(
-          `\nConfiguration saved to ${chalk.cyan(path.relative(process.cwd(), filePath))}`,
-        );
+        if (writeFile) {
+          await outputFile(filePath, content);
+          console.log(`\nConfiguration saved to ${chalk.cyan(path.relative(cwd, filePath))}`);
+        }
       }
     }),
   );
@@ -168,13 +179,17 @@ program
   .description('Fetch content objects')
   .option('-p, --preview', 'Fetch with preview mode')
   .option('-v, --verbose', 'Verbose output')
+  .option(
+    '--config <configFile>',
+    'Use this configuration, overriding other config options if present',
+  )
+  .option('--cwd <directory>', 'Working directory. Defaults to process.cwd()')
   .option('--sync', 'cache sync data')
   .option('--query <query>', 'Query used to fetch contentful entries')
   .option('--ignore-errors', 'No error return code when transform has errors')
   .action(
     actionRunner(async (cmd) => {
       const config = await getConfig(parseFetchArgs(cmd || {}));
-      const verified = await askMissing(config);
       const cache = initializeCache(config);
 
       if (cmd.sync && cmd.query) {
@@ -192,7 +207,7 @@ program
         await cache.reset();
       }
 
-      prev = await run({ ...verified, sync: Boolean(cmd.sync) }, prev);
+      prev = await run({ ...config, sync: Boolean(cmd.sync) }, prev);
 
       if (cmd.sync) {
         await cache.setSyncState(prev);
@@ -205,6 +220,11 @@ program
   .description('Fetch content objects && watch for changes')
   .option('-p, --preview', 'Fetch with preview mode')
   .option('-v, --verbose', 'Verbose output')
+  .option(
+    '--config <configFile>',
+    'Use this configuration, overriding other config options if present',
+  )
+  .option('--cwd <directory>', 'Working directory. Defaults to process.cwd()')
   .option('--url <url>', 'Webhook url.\nCan also be set via environment variable CSSG_WEBHOOK_URL')
   .option('--no-cache', "Don't cache sync data")
   .option('--poll', 'Use polling (usefull when ngrok tunnel is not an option)')
@@ -217,9 +237,8 @@ program
   .action(
     actionRunner(async (cmd) => {
       const config = await getConfig(parseFetchArgs(cmd || {}));
-      const verified = await askMissing(config);
       const useCache = Boolean(cmd?.cache ?? true);
-      const cache = initializeCache(verified);
+      const cache = initializeCache(config);
 
       let prev: RunResult;
       if (useCache && cache.hasSyncState()) {
@@ -228,7 +247,7 @@ program
         await cache.reset();
       }
 
-      prev = await run({ ...verified, sync: true }, prev);
+      prev = await run({ ...config, sync: true }, prev);
       if (useCache) {
         await cache.setSyncState(prev);
       }
@@ -255,7 +274,7 @@ program
           setTimeout(
             () => {
               (async () => {
-                prev = await run({ ...verified, sync: true }, prev);
+                prev = await run({ ...config, sync: true }, prev);
                 if (useCache) {
                   await cache.setSyncState(prev);
                 }
